experience tested

+++ src/modules/experience/experience.service.ts
@@ -6,6 +6,7 @@ import { Promise } from 'mongoose';
 import { CreateExperienceDto } from './dto/create-experience.dto';
 import { ExperienceResponse } from './dto/response-experience.dto';
 import { UpdateExperienceDto } from './dto/update-experience.dto';
+import { DeleteExperienceResponseDto } from './dto/delete-experience.dto';
 @Injectable()
 export class ExperienceService {
     constructor(
@@ -13,26 +14,30 @@ export class ExperienceService {
     ){}
 
 
-    async create (createExperienceDto: CreateExperienceDto): Promise<ExperienceResponse[]> {
-        const experience = new this.experienceModel(createExperienceDto);
+    async create (userId: string, createExperienceDto: CreateExperienceDto): Promise<ExperienceResponse[]> {
+
+
+        const experience = new this.experienceModel({...createExperienceDto, userId});
         await experience.save();
 
-        return await this.experienceModel.find({userId: createExperienceDto.userId}).exec()
+        const temp = await this.experienceModel.find({userId: userId}).select("designation companyName location startDate endDate description").exec()
+
+        return temp;
     }
 
-    async update (id: string, updateExperienceDto: UpdateExperienceDto):Promise<ExperienceResponse[]> {
+    async update (id: string, userId:string, updateExperienceDto: UpdateExperienceDto):Promise<ExperienceResponse[]> {
 
-        if(!Types.ObjectId.isValid(id)){
+        if(!Types.ObjectId.isValid(id) || !Types.ObjectId.isValid(userId)){
             throw new BadRequestException("Invalid experience id")
         }
 
-        const updated = await this.experienceModel.findOneAndUpdate({_id: id, userId: updateExperienceDto.userId}, updateExperienceDto, {new:true, runValidators:true})
+        const updated = await this.experienceModel.findOneAndUpdate({_id: id, userId:userId}, updateExperienceDto, {new:true, runValidators:true})
 
         if(!updated){
             throw new NotFoundException("Item may not exist")
         }
 
-        return await this.experienceModel.find({userId: updateExperienceDto.userId}).exec()
+        return await this.experienceModel.find({userId: userId}).select("designation companyName location startDate endDate description").exec()
 
 
     }
@@ -41,16 +46,32 @@ export class ExperienceService {
         if(!Types.ObjectId.isValid(userId)){
             throw new BadRequestException("Invalid experience id")
         }
-        return await this.experienceModel.find({userId: userId}).exec()
+        return await this.experienceModel.find({userId: userId}).select("designation companyName location startDate endDate description").exec()
     }
 
-    async remove (id: string, userId: string): Promise<any> {
+    async remove (id: string, userId: string): Promise<DeleteExperienceResponseDto> {
         if(!Types.ObjectId.isValid(userId) || !Types.ObjectId.isValid(userId)){
             throw new BadRequestException("Invalid experience id")
         }
+        const isFound = await this.experienceModel.findOne({_id:id, userId:userId}).select("designation companyName location startDate endDate description").exec()
+
+        if(!isFound){
+            throw new NotFoundException("There may not have experience")
+        }
+
+
         const removed = await this.experienceModel.deleteOne({_id: id, userId}).exec()
 
-        return removed
+        if(!removed){
+            throw new Error("Something went wrong")
+        }
+
+        const remaining = await this.experienceModel.find({userId:userId}).select("designation companyName location startDate endDate description").exec()
+
+        return {
+            removed: isFound,
+            remaining: remaining
+        }
     }
 
 }
